Split `if` helper test into a separate file
